1.6 - Reworking internal VeinMining system
* Introduce `/veintoggle` command

Toggle the usage of the VeinMining system - per player

You can also toggle the system by using the preferred tool, sneaking, and right clicking on a block

(i.e. You can sneak and right click on a log with an axe to toggle the system, the log will __not__ be stripped!)

* Allowed for `stripped` logs to be vein-mined

Seemed reasonable to allow this now that the user can toggle the system individually.